% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{add.objectrecord}
\alias{add.objectrecord}
\title{Register task's output object.}
\usage{
add.objectrecord(metadata, name, path = NULL, compress = "xz")
}
\arguments{
\item{metadata}{already created metadata you wish to add ancestor to. You can create task's metadata from scratch with \code{\link{create.metadata}}.}

\item{name}{name of the object's name}

\item{path}{location of the cached file with the computed object. The path can
be absolute or relative to the metadata's path.}

\item{compress}{optional. You can specify the compression method of the resulting object.
When used \code{'xz'} compression, the computer will try to compress it with the
parallel \code{pxz} application, if it is available on the system.}
}
\value{
modified \code{metadata} argument that includes specified object's record.
}
\description{
If you run tasks not for side effects, but for creating R objects, you need
to formally register each of them.
}
\details{
Each registered object gets cached to its own, unique file after a script run.
It can also be specified as a dependency of another task.

After script run, all objects that are not registered will be removed from memory.

If the object's record with specified name already exists, it will be silently overwritten.
}
\seealso{
\code{\link{create.metadata}}, \code{\link{add.parent}}
}

