% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{add.parent}
\alias{add.parent}
\title{Adds parent (ancestor) record to the existing metadata.}
\usage{
add.parent(metadata, name = NULL, metadata.path, aliasname = NULL)
}
\arguments{
\item{metadata}{already created metadata you wish to add ancestor to. You can create task's metadata from scratch with \code{\link{create.metadata}}.}

\item{name}{name of the object's name in the parent task you with to import}

\item{metadata.path}{either full or relative to metadata path with the
ancestor's metadata file.}

\item{aliasname}{optional argument. If set it specifies alternate name of the imported
\code{name} object as it is used by the task's R code. This setting allows great
flexibility in naming task's output objects.}
}
\value{
modified \code{metadata} argument that includes specified parent's record.
}
\description{
You specify parent record when you need another task to be already
computed as dependency before starting with the current one.
}
\details{
If the parent's record with specified name already exists, it silently overwrites it.
}
\seealso{
\code{\link{create.metadata}}, \code{\link{add.objectrecord}}
}

