% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{add_source_file}
\alias{add_source_file}
\title{Add additional source file to the task.}
\usage{
add_source_file(metadata, filepath, code = NULL, flag.checksum = TRUE)
}
\arguments{
\item{metadata}{already created metadata you wish to add the source file to. You can create task's metadata from scratch with \code{\link{create.metadata}}.}

\item{filepath}{Path to the file, if the file is already existing (it can be relative to the task's path)}

\item{code}{Optional string with the contents of the file. If specified and the file does not already exist, the file will be created
with this contents.}
}
\value{
modified \code{metadata} argument that includes additional source file or NULL if error.
}
\description{
Add additional source file to the task.
}
\seealso{
\code{\link{create.metadata}}, \code{\link{add.parent}}
}

