% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{append_extra_code}
\alias{append_extra_code}
\title{Function that appends file path to the metadata. It does no checking, it simply
updates the data structure.}
\usage{
append_extra_code(metadata, filepath, flag.checksum)
}
\description{
Function that appends file path to the metadata. It does no checking, it simply
updates the data structure.
}

