% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{are.two.metadatas.equal}
\alias{are.two.metadatas.equal}
\title{Tests whether two task metadata's define the same task equal (not necessarily identical)}
\usage{
are.two.metadatas.equal(m1, m2)
}
\arguments{
\item{m1}{First metadata to compare}

\item{m2}{Second metadata to compare}
}
\value{
\code{TRUE} if identical, \code{FALSE} otherwise
}
\description{
Tests whether two task metadata's define the same task equal (not necessarily identical)
}

