% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digests.R
\name{calculate.code.digest}
\alias{calculate.code.digest}
\title{Calculates R code component of the task's metadata.}
\usage{
calculate.code.digest(metadata)
}
\arguments{
\item{metadata}{Metadata of the object}
}
\description{
If there is more than one line of code, it calculates digest for each line separatedly,
then digests the final digest of concatenation of all individual digests.
If there is only one line of code it returns digest of it.
}

