% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digests.R
\name{calculate_one_digest}
\alias{calculate_one_digest}
\title{Calculated MD5 digest of a signle source file.
I don't use tools::md5sum to make sure, that the digest is independent on newlines format (Windows/Linux)}
\usage{
calculate_one_digest(code)
}
\description{
Calculated MD5 digest of a signle source file.
I don't use tools::md5sum to make sure, that the digest is independent on newlines format (Windows/Linux)
}

