% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_object_dirty.R
\name{code_has_been_changed}
\alias{code_has_been_changed}
\title{Returns TRUE if code has been changed by external force, forcing us to do
recalculation.}
\usage{
code_has_been_changed(metadata)
}
\description{
Returns TRUE if code has been changed by external force, forcing us to do
recalculation.
}

