% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{create.lock.file}
\alias{create.lock.file}
\title{Creates lock file on a given path. Lock file is just a 0-length file.}
\usage{
create.lock.file(path)
}
\arguments{
\item{path}{Location of the lock file}
}
\description{
The file will have an extension getOption('lock.extension')
}

