% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create.metadata}
\alias{create.metadata}
\title{Creates task's metadata object.}
\usage{
create.metadata(code, metadata.path, flag.never.execute.parallel = FALSE)
}
\arguments{
\item{code}{character vector with code. Each element of the vector will be treated as separate line.}

\item{metadata.path}{path to the task's metadata. The metadata will not be saved with this command, but
the path information is nevertheless required, as it is a mandatory part of the task's metadata specification.}

\item{flag.never.execute.parallel}{if set, the task will not be executed in parallel. Defaults to \code{FALSE}.}
}
\value{
task's metadata object. One might want to complete object's creation with complementary functions
   \code{add.parent} and \code{add.object.record}
}
\description{
The function will create the object in memory, it will not be saved to disk.
}
\seealso{
\code{\link{add.parent}}, \code{\link{add.objectrecord}}
}

