% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_objects.R
\name{create.objects}
\alias{create.objects}
\title{Function that creates the object by calling execution of its script.}
\usage{
create.objects(metadata, objects.to.keep = NULL, objectaliases, metadata.path,
  flag.save.intermediate.objects = TRUE, flag.check.md5sum = TRUE,
  flag.save.in.background = TRUE, flag.forget.parents = TRUE,
  estimation.only = NULL)
}
\arguments{
\item{metadata}{Metadata}

\item{objects.to.keep}{List of objects, which will be needed by later tasks and should not be removed from memory}

\item{objectaliases}{Alternate names created objects}

\item{metadata.path}{metadata.path}

\item{flag.save.intermediate.objects}{If true, all created objects will be cached. Default: \code{TRUE}}

\item{flag.check.md5sum}{Gets forwarded to \code{load.objects} function}

\item{flag.save.in.background}{If true, the process of saving the objects gets forked into another
process and done in parallel in background to speed-up computations. Default: \code{TRUE}}

\item{flag.forget.parents}{If set, all required prerequisite objects will get removed from the memory,
after the script is run. Default: \code{TRUE}}

\item{estimation.only}{If \code{NULL} the function is run for its side effects. If not null, it makes no side effects
and instead it returns detailed trace of the computations it is going to perform. This information is gathered
by high-level function \code{metadata.dump}}
}
\value{
If \code{is.null(estiation.only)} it returns boolean, where \code{TRUE} means success,
and \code{FALSE} means failure.
Otherwise it returns updated run statistics
}
\description{
Function that creates the object by calling execution of its script.
}

