% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.objectrecords}
\alias{get.objectrecords}
\title{Returns specific subset of objectrecords}
\usage{
get.objectrecords(metadata, objnames)
}
\arguments{
\item{metadata}{Task's metadata}

\item{objnames}{Character vector with names of objects for which we want to
get objectrecords}
}
\description{
Returns specific subset of objectrecords
}

