% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digests.R
\name{get_coding_files}
\alias{get_coding_files}
\title{Returns all files with code in canonical order:
first the main file,
then all the auxilary files sorted alphabetically including paths}
\usage{
get_coding_files(metadata, flag_expand_paths)
}
\description{
Returns all files with code in canonical order:
first the main file,
then all the auxilary files sorted alphabetically including paths
}

