% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load.and.validate.parents}
\alias{load.and.validate.parents}
\title{Recursively loads and validates all parent records for the given metadata}
\usage{
load.and.validate.parents(metadata, flag.check.md5 = FALSE,
  estimation.only = NULL)
}
\arguments{
\item{metadata}{Metadata for which we want to load all parents (dependencies)}

\item{flag.check.md5}{This parameter gets forwarded to \code{load.object.from.disk}}

\item{estimation.only}{If set, the function performs a dry run, with no parents get loaded,
but extra information gets returned.}
}
\value{
If \code{estimation.only=NULL}, it returns boolean in which \code{TRUE} means success
  and \code{FALSE} - failure.

  If \code{estimation.only!=NULL} it returns list used by \code{metadata_dump}
}
\description{
It is a relatively high-level function that participates in recursive loading of the
tasks. It is called to make sure, that the immidiate parents are valid and loaded into
memory under the correct aliasnames.
}
\details{
The function operates in two modes: one with \code{estimation.only=NULL}, where the function
is run for its side-effects which result in loading all the necessary prerequisite objects
for the task. Otherwise, if \code{estimation.only!=NULL}, the function makes no side effects, but instead
it returns detailed and recursive trace of all tasks, that need to be calculated in order
to satisfy the dependencies.
}

