% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_load.R
\name{load.metadata}
\alias{load.metadata}
\title{Loads the task's metadata object from disk}
\usage{
load.metadata(metadata.path)
}
\arguments{
\item{metadata.path}{Full path to the metadata.}
}
\value{
metadata object.
}
\description{
Loads the task's metadata object from disk
}

