% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load.object.from.disk}
\alias{load.object.from.disk}
\title{Loads one or more tasks' objects from disk cache.}
\usage{
load.object.from.disk(metadata, objectrecord, aliasname = NULL,
  flag.dont.load = FALSE, flag.check.md5 = TRUE)
}
\arguments{
\item{metadata}{Task's metadata}

\item{objectrecord}{Objectrecord item that corresponds to the cached file we want to load}

\item{aliasname}{If specified, it overrides the name, under which the loaded object will be available to R.}

\item{flag.dont.load}{If \code{TRUE}, no object ever gets loaded, and \code{aliasname} parameter is ignored.
It allows to check the availability of the cached object instead of loading it}

\item{flag.check.md5}{If set, it also checks MD5 hash of the file. For large files it can be relatively long
process.}
}
\value{
Character string that describes the reason, why the object could not be loaded, or
  'OK' string to indicate that the object was or can be loaded.
}
\description{
Object gets loaded only if its file attributes precisely matche those saved in metadata.
}
\details{
Full list of file properties  that have to match in order for being loaded:

\enumerate{
 \item The file must not be locked by another process - i.e. there must
   be no file prefixed with .lock present in the directory
 \item Objectrecord must have metrics with which we can compare the file properties
 \item File size must match
 \item Modification time (mtime) must match
 \item If \code{flag.check.md5==TRUE}, file MD5 hash must match
}
}

