% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dependency.R
\name{load.object}
\alias{get.object}
\alias{load.object}
\title{Gets the object described by the task}
\usage{
load.object(metadata.path = NULL, metadata = NULL, objectnames = NULL,
  flag.save.intermediate.objects = TRUE, flag.check.md5sum = TRUE,
  flag.save.in.background = TRUE, flag.check.object.digest = TRUE)

get.object(metadata.path = NULL, metadata = NULL, objectname = NULL,
  flag.save.intermediate.objects = TRUE, flag.check.md5sum = TRUE,
  flag.save.in.background = TRUE, flag.check.object.digest = TRUE)
}
\arguments{
\item{metadata.path}{Path to the task's metadata, that describe the task. Alternatively user can supply}

\item{metadata}{The metadata object itself. It makes sure, that this metadata is saved to disk.}

\item{objectnames}{List of objects to return. If omited, all objects registered in this metadata will be returned.}

\item{flag.save.intermediate.objects}{If true, all intermediate objects will be saved, to speed-up future retrieval.
Default: \code{TRUE}.}

\item{flag.check.md5sum}{This flag is forwarded to \code{save.object} function and is in effect only if
\code{flag.save.intermediate.objects==TRUE}. If set, the \code{save.object} would record file's MD5 hash to the
metadata.}

\item{flag.save.in.background}{If set, objects will be saved in background, leaving the foreground thread
ready for other things. Default: \code{TRUE}. Ignored, if \code{flag.save.intermediate.objects==FALSE}.}

\item{flag.check.object.digest}{If set, MD5 signature of all created R objects will be computed and
stored in metadata. It will be used to further verify authenticity of the object, when retrieved from memory.}

\item{objectname}{Object to return.}
}
\value{
\code{load.object} returns updated metadata on success, and \code{NULL} on failure.
  \code{get.object} returns the object itself, if there is only one object to return,
  a named list of objects, if requested to return more than one object, or \code{NULL} on failure.
}
\description{
This function first tries to grap the object from R memory, if already present.
}
\details{
If not, it tries to get the object from saved disk cache.

Both functions \code{load.object} load the object into the global environment,
\code{get.object} also returns the object(s).

If no cache is available, it prepares all prerequisites for the script that runs the task,
and then it runs the script itself.

It makes sure, that all results are saved to speedup future retrievals.
}
\section{Functions}{
\itemize{
\item \code{get.object}: Gets the object by its metadata.
}}

