% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dependency.R
\name{load.objects.by.metadata}
\alias{load.objects.by.metadata}
\title{Hi level function, that reads objectnames under the names \code{aliasnames}.}
\usage{
load.objects.by.metadata(metadata, metadata.path, objectnames,
  aliasnames = NULL, flag.save.intermediate.objects = TRUE,
  flag.check.md5sum = TRUE, flag.save.in.background = TRUE,
  flag.check.object.digest = TRUE, flag.estimate.only = FALSE)
}
\arguments{
\item{metadata}{The metadata object itself. It makes sure, that this metadata is saved to disk.}

\item{metadata.path}{Path to the task's metadata, that describe the task. Alternatively user can supply}

\item{objectnames}{List of objects to return. If omited, all objects registered in this metadata will be returned.}

\item{aliasnames}{List of new names for the created objects.}

\item{flag.save.intermediate.objects}{If true, all intermediate objects will be saved, to speed-up future retrieval.
Default: \code{TRUE}.}

\item{flag.check.md5sum}{This flag is forwarded to \code{save.object} function and is in effect only if
\code{flag.save.intermediate.objects==TRUE}. If set, the \code{save.object} would record file's MD5 hash to the
metadata.}

\item{flag.save.in.background}{If set, objects will be saved in background, leaving the foreground thread
ready for other things. Default: \code{TRUE}. Ignored, if \code{flag.save.intermediate.objects==FALSE}.}

\item{flag.check.object.digest}{If set, MD5 signature of all created R objects will be computed and
stored in metadata. It will be used to further verify authenticity of the object, when retrieved from memory.}

\item{flag.estimate.only}{If set, dry run will be performed, during which the
run metrics will be gathered and returned for use of \code{metadata_dump}.}
}
\value{
Returns updated metadata on success, and \code{NULL} on failure.
}
\description{
If \code{flag.estimate.only} is present, it returns the detailed execution tree and historic execution metrics
 instead of executing the scripts.
}

