% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digests.R
\name{metadata.digest}
\alias{metadata.digest}
\title{Returns digest of the task's metadata.}
\usage{
metadata.digest(metadata)
}
\arguments{
\item{metadata}{task's metadata object, based on which calculate the digest.

The digest compirses the following items:

\enumerate{
\item names and aliasnames of all parents
\item code
\item names of all exported symbols
}
Everything else is ignored.}
}
\value{
string representing MD5 digest of the object in lowercase.
}
\description{
Digests are used in many places for easy identification of equivalent tasks' metadatas.
}
\details{
The digest is calculated on the task's essential properties:
\itemize{
 \item \strong{parent records}.
   \describe{
     \item{\code{path}}{path to the ancestor's task metadata}
     \item{\code{name}}{name of the R object calculated by the ancestor}
     \item{\code{aliasname}}{optional. Specifies alternate name of the object for our R script}
   }
 \item \strong{code}.
   All source code used by our task
 \item \strong{output object records}
   \describe{
     \item{\code{name}}{name of the output object}
   }
}
}

