% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dump.R
\name{metadata.dump}
\alias{metadata.dump}
\title{Metadata in tabular form}
\usage{
metadata.dump(metadata.path, id = NULL)
}
\arguments{
\item{metadata.path}{Path to task's metadata object}

\item{id}{Arbitrary ID if this metadata, used as a root for all descendants (ancestors).
For details see description.}
}
\value{
Returns list of 4 relational tables that fully describe the metadata of the task.
It is should be used, when the tool used to communicate with R does not support lists.

Records in the main \code{metadata} table are identified by the \code{ID} field, which
contains arbitrary number that distinguish different tasks between each other.
The ID of the object loaded from the function argument can be set arbitrarily as
\code{id} parameter for ease of parsing the tables with external tools.
\strong{\code{metadata}}
\describe{
  \item{\code{id}}{task's unique ID}
  \item{\code{flags}}{integer with the following binary flags:
     \enumerate{
       \item \strong{1} - All files with code exist on hard drive
       \item \strong{2} - When set, never execute the task in parallel
       \item \strong{4} - Don't use cached value, force recalculation of this task
     }}
  \item{\code{path}}{full path to the task metada file}
  \item{\code{codepath}}{full path to the R source code}
  \item{\code{metadatadigest}}{md5 digest of the task's metadata.
     The digest is calculated from three components:
     \enumerate{
       \item digest of parents
       \item digest of exported objects (\code{objectrecords})
       \item digest of all source code files
     }
     Digest is being used to check wether two tasks' metadatas describe equivalent objects}
  \item{\code{codedigest}}{md5 digest of all source files. It is a component of \code{metadatadigest}}
  \item{\code{code}}{R source code of the task.  If the task needs other code to work,
   this code needs to be read separately.}
}
\strong{\code{objectrecords}}
\describe{
  \item{\code{parentid}}{task ID, that links each row from this table with table \code{metadata}
   in 1-to-many relationship}
  \item{\code{flags}}{integer with the following binary flags:
     \enumerate{
       \item \strong{1} - The object is already in R's memory
       \item \strong{2} - The cached object exists, but maybe not update.
       \item \strong{4} - The cached object exists, and is update and ready for being loaded without re-computation
     }}
  \item{\code{path}}{full path to the cached object}
  \item{\code{name}}{R name of the object}
  \item{\code{compress}}{string describing compression level used when saving cache of the object}
  The following items exist only if the task was performed:
  \item{\code{size}}{size of the R object, when loaded into R's memory in bytes}
  \item{\code{objectdigest}}{digest of the R object, when loaded into R's memory}
  The following items exist only if the task was cached:
  \item{\code{mtime}}{modification time of the cached object or NULL if the cached object doesn't exist }
  \item{\code{filedigest}}{md5 digest of the cached file if the file exists. It can be used to check whether actual
    cache is consistent with the task's metafile description}
  \item{\code{filesize}}{size of the cached file in bytes}
}
\strong{\code{parents}}
\describe{
  \item{\code{parentid}}{task's metadata ID, that links each row from this table with table \code{metadata}
   in 1-to-many relationship}
  \item{\code{metadatadigest}}{md5 digest of the parent task's metadata}
  \item{\code{name}}{original name of the imported parent's object}
  \item{\code{aliasname}}{if present, it gives R's name of the parent's object that is
   being used in the task's R script. This field gives flexibility in naming objects across tasks}
  \item{\code{path}}{path to the parent's task's metadata}
 }
\strong{\code{timecosts}}

Each row describe statistics of each recalculation of this task.
\describe{
  \item{\code{parentid}}{task's metadata ID, that links each row from this table with table \code{metadata}
   in 1-to-many relationship}
 \item{\code{walltime}}{Wall time of the task calculation}
 \item{\code{cputime}}{CPU user time of the task calculation. Can be greater than \code{walltime} for parallel tasks,
   and less than \code{walltime} for disk-bound tasks.}
 \item{\code{systemtime}}{CPU kernel time of the task calculation. }
 \item{\code{cpumodel}}{Name of the CPU on the machine, where the task was calculated.
   We don't support machines with heterogenous CPUs}
 \item{\code{corecount}}{Number of CPU cores (\emph{virtual cores are not counted)}}
 \item{\code{virtualcorecount}}{Number of virtual cores (including Intel's hyperthreading virtual cores)}
 \item{\code{busycpus}}{Number of busy virtual cores just before the task was started.
   It gives a measure of system load when computing the task.}
 \item{\code{membefore}}{Amount of free system memory before task was started}
 \item{\code{memafter}}{Amount of free system memory immidiately after calculation and \code{gc()} .
   Memory fragmentation, leakage, and resulting object sizes all contribute to this memory}
}
}
\description{
Metadata in tabular form
}

