% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{metadata.objects.size}
\alias{metadata.objects.size}
\title{Returns total size of all objects or NA}
\usage{
metadata.objects.size(metadata)
}
\arguments{
\item{metadata}{Task's metadata}
}
\description{
Returns total size of all objects or NA
}

