% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digests.R
\name{objects.digest}
\alias{objects.digest}
\title{Returns object's digest for testing equivalence of metadataas
Digest is solely based on names of the objects:}
\usage{
objects.digest(metadata)
}
\arguments{
\item{metadata}{Task's metadata}
}
\description{
Returns object's digest for testing equivalence of metadataas
Digest is solely based on names of the objects:
}

