% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digests.R
\name{parents.digest}
\alias{parents.digest}
\title{Returns digest of object's parent information.
This include:
1. name of the imported object
2. aliasname of the imported object as it is used by us}
\usage{
parents.digest(metadata)
}
\description{
Returns digest of object's parent information.
This include:
1. name of the imported object
2. aliasname of the imported object as it is used by us
}

