% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{save.object}
\alias{save.object}
\title{Saves the object to disk.}
\usage{
save.object(metadata, objectrecord, flag.check.md5sum = TRUE,
  flag.forget.object = FALSE)
}
\arguments{
\item{metadata}{Metadata that describes the task for which the object is being saved}

\item{objectrecord}{Object record that describes the object being saved.}

\item{flag.check.md5sum}{If set, MD5 hash of the saved file will be calculated and included in the
metadata. Default \code{TRUE}}

\item{flag.forget.object}{If set, the function will remove the object from memory and perform garbage
collection as soon, as possible.}

\item{metadata.path}{Unused}
}
\value{
'OK' if success, or string that describe an error if failed.
}
\description{
It saves the object as a \code{.rds} file using internal R format.
}
\details{
For space efficiency it compresses the object with method specified by \code{$compress} method.

To speedup rather slow \code{xz} compression, the program tries to take advantage of \code{pxz} utility
(Parallel PXZ) if it is available.

It also updates task's metadata file with numerous computation statistics and file attributes.
}

