% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort.parentrecords}
\alias{sort.parentrecords}
\title{Returns sorted parentrecords.}
\usage{
\method{sort}{parentrecords}(metadata)
}
\description{
@param metadata Metadata object
 @return  Sorted lists of items. Each item corresponds to one imported task and is a list of elements:
 \describe{
 \item{\code{path}}{path to the metadata}
 \item{\code{names}}{character vector with names of exported objects from this metadata}
 \item{\code{aliasnames}}{optional: character vector of the names used by our script}
 }
}

