% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{unload.objects}
\alias{unload.objects}
\title{Sets the objects to NULL, effectively unloading them.}
\usage{
unload.objects(parentrecords)
}
\arguments{
\item{parentrecords}{Parentrecords that describe objects to unload.}
}
\description{
Sets the objects to NULL, effectively unloading them.
}

